+++ Runtime/Singleton/Singleton.cs
@@ -1,0 +1,68 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace HelloWorld.Utils
+{
+    public class Singleton<T> : MonoBehaviour where T : MonoBehaviour
+    {
+        private static T _instance;
+        private static readonly object _lock = new object();
+        private static bool _applicationIsQuitting = false;
+
+        public static T Instance
+        {
+            get
+            {
+                if (_applicationIsQuitting)
+                {
+                    Debug.LogWarning($"[Singleton] Instance of {typeof(T)} is already destroyed. Returning null.");
+                    return null;
+                }
+
+                lock (_lock)
+                {
+                    if (_instance == null)
+                    {
+                        _instance = FindObjectOfType<T>();
+
+                        if (FindObjectsOfType<T>().Length > 1)
+                        {
+                            Debug.LogError($"[Singleton] There are multiple instances of {typeof(T)} in the scene. This is not allowed.");
+                            return _instance;
+                        }
+
+                        if (_instance == null)
+                        {
+                            GameObject singletonObject = new GameObject($"(Singleton) {typeof(T)}");
+                            _instance = singletonObject.AddComponent<T>();
+                            DontDestroyOnLoad(singletonObject);
+                            Debug.Log($"[Singleton] Created instance of {typeof(T)}");
+                        }
+                    }
+
+                    return _instance;
+                }
+            }
+        }
+
+        protected virtual void Awake()
+        {
+            if (_instance == null)
+            {
+                _instance = this as T;
+                DontDestroyOnLoad(gameObject);
+            }
+            else if (_instance != this)
+            {
+                Debug.LogWarning($"[Singleton] Destroying duplicate instance of {typeof(T)} attached to {gameObject.name}.");
+                Destroy(gameObject);
+            }
+        }
+
+        private void OnApplicationQuit()
+        {
+            _applicationIsQuitting = true;
+        }
+    }
+}