+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# CHANGELOG
+
 ## [1.0.0] - 2025-06-17
+
 ### First Release
-- Add utils
+
+- Add singleton
+- Add localization
+- Add horizontal selector
+- Add scene loader